z_requirements.txt
# pip install -r z_requirements.txt
python-dotenv
requests
urllib3
pynacl
cryptography
aiohttp
asyncio
tk
